# gpaserveisWeb/src/main/resources/app/config/gpaserveis-swagger-rest-api.yaml
{
  "swagger": "2.0",
  "info": {
    "description": "Capa de serveis, funcions i accions",
    "version": "1.0.0",
    "title": "Capa de serveis, funcions i accions",
    "contact": {
      "name": "Sergio Corras",
      "email": "sergio.corras@iecisa.com"
    },
    "license": {
      "name": "Llicència IMI"
    }
  },
  "host": "rpa-int.icp.ajuntament.bcn",
  "basePath": "/services/gpaserveis/rest",
...
  "paths": {
    "/serveis/portal/expedients": {
      "get": {
        "tags": [
          "Funcions de consulta al repositori de dades d'expedients",
          "Serveis Portal API"
        ],
        "summary": "Cerca d'expedients",
        "operationId": "cercaExpedientsUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "codiExpedient",
            "in": "query",
            "description": "Filtra expedients per codi",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "codiProcediment",
            "in": "query",
            "description": "Filtra expedients per codi del procediment",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "dataPresentacioFi",
            "in": "query",
            "description": "Filtra expedients per data de presentació",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "dataPresentacioInici",
            "in": "query",
            "description": "Filtra expedients per data de presentació",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "estat",
            "in": "query",
            "description": "Filtra expedients per estat",
            "required": false,
            "type": "string",
            "allowEmptyValue": false,
            "enum": [
              "EN_PREPARACIO",
              "EN_REVISIO",
              "PENDENT_SUBSANACIO",
              "EN_TRAMITACIO",
              "PENDENT_ALEGACIONS",
              "PENDENT_INFORMES",
              "PROPOSAT_RESOLUCIO",
              "RESOLT",
              "TANCAT"
            ]
          },
          {
            "name": "nifSolicitant",
            "in": "query",
            "description": "Filtra expedients d'un sol·licitant",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "numeroPagina",
            "in": "query",
            "description": "Indicarà en quina pàgina hauria de començar els resultats demanats en una cerca",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "resultatsPerPagina",
            "in": "query",
            "description": "Indicarà el número de resultats per pàgina",
            "required": false,
            "type": "integer",
            "default": 20,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "unitatGestora",
            "in": "query",
            "description": "Filtra expedients per Unitat Gestora",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaCercaProcediments"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "consulta": "Perfil usuari consulta"
        }
      },
      "post": {
        "tags": [
          "Funcions d'execució d'accions",
          "Serveis Portal API"
        ],
        "summary": "Crear una sol·licitud d'un expedient",
        "operationId": "crearSolicitudExpedientUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "solicitudExpedient",
            "description": "Identificador de l'expedient",
            "required": false,
            "schema": {
              "$ref": "#/definitions/SolicitudCrear"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaCrearSolicitud"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "consulta": "Perfil usuari consulta"
        }
      }
    },
    "/serveis/portal/expedients/{idExpedient}": {
      "get": {
        "tags": [
          "Funcions de consulta al repositori de dades d'expedients",
          "Serveis Portal API"
        ],
        "summary": "Consultar les dades de l'expedient",
        "operationId": "consultarDadesExpedientUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idExpedient",
            "in": "path",
            "description": "Identificador de l'expedient",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaConsultaExpedients"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "consulta": "Perfil usuari consulta"
        }
      },
      "put": {
        "tags": [
          "Funcions d'execució d'accions",
          "Serveis Portal API"
        ],
        "summary": "Actualitzar dades de la sol·licitud de l'expedient",
        "operationId": "actualitzarSolicitudExpedientUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idExpedient",
            "in": "path",
            "description": "Identificador de l'expedient",
            "required": true,
            "type": "number"
          },
          {
            "in": "body",
            "name": "solicitudExpedient",
            "description": "Identificador de l'expedient",
            "required": false,
            "schema": {
              "$ref": "#/definitions/SolicitudActualitzar"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaActualitzarSolicitud"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "gestor": "Perfil usuari gestor"
        }
      }
    },
    "/serveis/portal/expedients/{idExpedient}/obrir": {
      "post": {
        "tags": [
          "Funcions d'execució d'accions",
          "Serveis Portal API"
        ],
        "summary": "Obrir la sol·licitud de l'expedient",
        "operationId": "obrirSolicitudExpedientUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idExpedient",
            "in": "path",
            "description": "Identificador de l'expedient",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaObrirSolicitud"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "gestor": "Perfil usuari gestor"
        }
      }
    },
    "/serveis/portal/expedients/{idExpedient}/registre": {
      "post": {
        "tags": [
          "Funcions d'execució d'accions",
          "Serveis Portal API"
        ],
        "summary": "Registrar la sol·licitud de l'expedient",
        "operationId": "registrarSolicitudExpedientUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idExpedient",
            "in": "path",
            "description": "Identificador de l'expedient",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaRegistrarSolicitud"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "gestor": "Perfil usuari gestor"
        }
      }
    },
    "/serveis/portal/procediments": {
      "get": {
        "tags": [
          "Funcions d'integració amb RPA",
          "Serveis Portal API"
        ],
        "summary": "Cerca de procediments",
        "operationId": "cercaProcedimentsUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "actiu",
            "in": "query",
            "description": "Filtra aquells procediments entre els que tinguin el període de sol·licitud obert i els que no",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "codi",
            "in": "query",
            "description": "Filtra procediments per codi",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "descripcio",
            "in": "query",
            "description": "Filtra procediments per descripció",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "estat",
            "in": "query",
            "description": "Filtra procediments per estat",
            "required": false,
            "type": "string",
            "allowEmptyValue": false,
            "enum": [
              "EN_ELABORACIO",
              "PUBLICAT",
              "FINALITZAT"
            ]
          },
          {
            "name": "inici",
            "in": "query",
            "description": "Filtra procediments per tipus d'inici",
            "required": false,
            "type": "string",
            "allowEmptyValue": false,
            "enum": [
              "OFICI",
              "SOLICITUD",
              "ADMINISTRACIO",
              "ALTRES"
            ]
          },
          {
            "name": "intern",
            "in": "query",
            "description": "Filtra procediments interns",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "numeroPagina",
            "in": "query",
            "description": "Indicarà en quina pàgina hauria de començar els resultats demanats en una cerca",
            "required": false,
            "type": "integer",
            "default": 1,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "resultatsPerPagina",
            "in": "query",
            "description": "Indicarà el número de resultats per pàgina",
            "required": false,
            "type": "integer",
            "default": 20,
            "format": "int32",
            "allowEmptyValue": false
          },
          {
            "name": "ugo",
            "in": "query",
            "description": "Filtra procediments per Unitat Gestora Operativa",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "ugr",
            "in": "query",
            "description": "Filtra procediments per Unitat Gestora Responsable",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          },
          {
            "name": "versio",
            "in": "query",
            "description": "Filtra procediments per versió",
            "required": false,
            "type": "string",
            "allowEmptyValue": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaCercaProcediments"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "consulta": "Perfil usuari consulta"
        }
      }
    },
    "/serveis/portal/procediments/{idProcediment}": {
      "get": {
        "tags": [
          "Funcions d'integració amb RPA",
          "Serveis Portal API"
        ],
        "summary": "Consultar les dades del procediment",
        "operationId": "consultarDadesProcedimentUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idProcediment",
            "in": "path",
            "description": "Identificador del procediment",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaConsultaProcediments"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "consulta": "Perfil usuari consulta"
        }
      }
    },
    "/serveis/portal/procediments/{idProcediment}/tramits/{idTramit}/atributs": {
      "get": {
        "tags": [
          "Funcions d'integració amb RPA",
          "Serveis Portal API"
        ],
        "summary": "Consultar les dades d'operació del tràmit",
        "operationId": "consultarDadesOperacioProcedimentUsingGET",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "idProcediment",
            "in": "path",
            "description": "Identificador del procediment",
            "required": true,
            "type": "number"
          },
          {
            "name": "idTramit",
            "in": "path",
            "description": "Identificador del tràmit",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RespostaDadesOperacio"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false,
        "x-imi-roles": {
          "consulta": "Perfil usuari consulta"
        }
      }
    },
...
  },
  "definitions": {
...
    "RespostaActualitzarSolicitud": {
      "type": "object",
      "properties": {
        "expedient": {
          "description": "Expedient actualitzat en estat esborrany",
          "$ref": "#/definitions/RespostaExpedientCrear"
        },
        "resultat": {
          "description": "Resultat de l'operació",
          "$ref": "#/definitions/UnitatGestora"
        }
      },
      "title": "RespostaActualitzarSolicitud"
    },
    "RespostaCercaProcediments": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ExpedientsCerca"
          }
        },
        "paginacio": {
          "$ref": "#/definitions/Paginacio"
        }
      },
      "title": "RespostaCercaProcediments"
    },
...
    "RespostaConsultaExpedients": {
      "type": "object",
      "properties": {
        "expedient": {
          "$ref": "#/definitions/ExpedientsConsulta"
        }
      },
      "title": "RespostaConsultaExpedients"
    },
    "RespostaConsultaProcediments": {
      "type": "object",
      "properties": {
        "procediment": {
          "$ref": "#/definitions/ProcedimentConsulta"
        }
      },
      "title": "RespostaConsultaProcediments"
    },
    "RespostaCrearSolicitud": {
      "type": "object",
      "properties": {
        "expedient": {
          "description": "Expedient creat en estat inicial",
          "$ref": "#/definitions/RespostaExpedientCrear"
        },
        "resultat": {
          "description": "Resultat de l'operació",
          "$ref": "#/definitions/UnitatGestora"
        }
      },
      "title": "RespostaCrearSolicitud"
    },
...
    "SolicitudCrear": {
      "type": "object",
      "properties": {
        "codiProcediment": {
          "type": "string",
          "description": "Codi del procediment"
        },
        "codiUnitatGestora": {
          "type": "string",
          "description": "Codi de la Unitat Gestora"
        },
        "inici": {
          "type": "string",
          "description": "Tipus d'inici de l'expedient",
          "enum": [
            "OFICI",
            "SOLICITUD",
            "ADMINISTRACIO",
            "ALTRES"
          ]
        },
        "representant": {
          "description": "Persona representant",
          "$ref": "#/definitions/Persona"
        },
        "solicitant": {
          "description": "Persona sol·licitant",
          "$ref": "#/definitions/Persona"
        },
        "versioProcediment": {
          "type": "string",
          "description": "Versió del procediment"
        }
      },
      "title": "SolicitudCrear"
    },
...
  }
}